nass-853: remove ajv
