+++ packages/desktop/app/views/administration/reports/EditReportView.js
@@ -1,10 +1,9 @@
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import React, { useState } from 'react';
+import { useQueryClient } from '@tanstack/react-query';
+import React from 'react';
 import { toast } from 'react-toastify';
 import styled from 'styled-components';
 import { Box } from '@material-ui/core';
 import { useApi } from '../../../api';
-import { ReportTable, VersionTable } from './ReportTables';
 import { useAuth } from '../../../contexts/Auth';
 import { OutlinedButton } from '../../../components';
 import { Colors } from '../../../constants';
@@ -13,15 +12,6 @@ import { ReportEditor } from './ReportEditor';
 
 const Container = styled.div`
   padding: 20px;
-`;
-
-const FlexContainer = styled(Container)`
-  display: flex;
-  align-items: flex-start;
-`;
-
-const VersionsTableContainer = styled.div`
-  margin-left: 20px;
 `;
 
 const StyledButton = styled(OutlinedButton)`
@@ -40,11 +30,11 @@ const getInitialValues = (version, report) => {
     query,
     status,
     ...options,
-    dataSources: dataSources.join(', '),
+    dataSources: dataSources?.join(', '),
   };
 };
 
-const VersionEditorView = ({ report, version, setVersion, onBack }) => {
+export const EditReportView = ({ report, version, setVersion, onBack }) => {
   const api = useApi();
   const { currentUser } = useAuth();
   const queryClient = useQueryClient();
@@ -86,60 +76,5 @@ const VersionEditorView = ({ report, version, setVersion, onBack }) => {
         initialValues={getInitialValues(version, report)}
       />
     </Container>
-  );
-};
-
-export const EditReportView = () => {
-  const [report, setReport] = useState(null);
-  const [version, setVersion] = useState(null);
-  const api = useApi();
-
-  const { data: reportData = [], isLoading: isReportLoading, error: reportError } = useQuery(
-    ['reportList'],
-    () => api.get('admin/reports'),
-  );
-
-  const areVersionsEnabled = !!report?.id;
-  const { data: versionData, isLoading: areVersionsLoading, error: versionsError } = useQuery(
-    ['reportVersions', report?.id],
-    () => api.get(`admin/reports/${report?.id}/versions`),
-    {
-      enabled: areVersionsEnabled,
-    },
-  );
-
-  const handleBack = () => setVersion(null);
-
-  return (
-    <>
-      {version ? (
-        <VersionEditorView
-          report={report}
-          version={version}
-          onBack={handleBack}
-          setVersion={setVersion}
-        />
-      ) : (
-        <FlexContainer>
-          <ReportTable
-            data={reportData}
-            selected={report?.id}
-            onRowClick={setReport}
-            loading={isReportLoading}
-            error={reportError?.message}
-          />
-          {versionData && (
-            <VersionsTableContainer>
-              <VersionTable
-                data={versionData}
-                loading={areVersionsEnabled && areVersionsLoading}
-                error={versionsError?.message}
-                onRowClick={setVersion}
-              />
-            </VersionsTableContainer>
-          )}
-        </FlexContainer>
-      )}
-    </>
   );
 };
