+++ packages/desktop/package.json
@@ -191,7 +191,6 @@
     "webpack-merge": "^4.1.3"
   },
   "dependencies": {
-    "ajv": "^7.2.4",
     "@babel/core": "^7.17.12",
     "@babel/plugin-proposal-class-properties": "^7.17.12",
     "@babel/plugin-proposal-export-default-from": "^7.17.12",
